pyproject.toml
[tool.poetry]
name = "dados-publicos"
version = "0.1.0"
description = ""
authors = ["Andre Luiz Braga Mallmann <andrebmallmann@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
google-cloud-storage = "^2.14.0"
google-auth = "^2.28.1"
google-cloud-bigquery = "^3.17.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.dagster]
